Improved makefile

+++ scripts/test_down.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 set -e
 
-docker kill talon_mariadb talon_postgres
-echo "containers stopped"
+if [[ -n $(docker ps -q -f name=talon_) ]]; then
+	docker kill talon_mariadb talon_postgres
+	echo "containers stopped"
+fi
